feat(modules): remove function names + add `numToBool` submodule + move `removeDuplicates` module to `util`
