+++ src/modules.js
@@ -7,9 +7,9 @@ module.exports = {
         "boolToNum": require("./modules/util/boolToNum"),
         "boolToStr": require("./modules/util/boolToStr"),
         "numToStr": require("./modules/util/numToStr"),
+        "removeDuplicates": require("./modules/util/removeDuplicates"),
         "strToBool": require("./modules/util/strToBool"),
         "strToNum": require("./modules/util/strToNum")
     },
-    "removeDuplicates": require("./modules/removeDuplicates"),
     "uuid": require("./modules/uuid")
 }
