+++ src/modules/log.js
@@ -1,4 +1,4 @@
-module.exports = function log(input) {
+module.exports = function (input) {
     if(!input) {
         throw new Error("No input specified");
     }
