+++ src/modules/util/boolToStr.js
@@ -1,4 +1,4 @@
-module.exports = function boolToStr(boolean) {
+module.exports = function (boolean) {
     if(typeof boolean !== "boolean") throw new Error("No boolean provided");
 
     return boolean.toString();
