+++ src/modules/util/numToStr.js
@@ -1,5 +1,5 @@
-module.exports = function numToStr(number) {
-    if(!number || (typeof number !== "number")) throw new Error("No number provided");
+module.exports = function (number) {
+    if(typeof number !== "number") throw new Error("No number provided");
 
     return number.toString();
 }
