+++ src/modules/util/strToBool.js
@@ -1,4 +1,4 @@
-module.exports = function strToBool(string) {
+module.exports = function (string) {
     if(!string || (typeof string !== "string")) throw new Error("No string provided");
 
     if(string !== "true" && string !== "false") throw new Error("No valid string provided");
